testsuite/gdb.dwarf2: avoid dead code in dw2-inline-with-lexical-scope.c
The test in gdb.dwarf2/dw2-inline-with-lexical-scope.c fails with icc.
The reason is, icc did not emit code for a dead statement, which in
turn caused some labels to be collapsed.  Fix this by replacing the
dead code with assignment to a global value.  The statement itself
does not change the test scenario.

Also fix a whitespacing problem around an assignment operator.

gdb/testsuite/ChangeLog:
2021-05-21  Tankut Baris Aktemur  <tankut.baris.aktemur@intel.com>

	* gdb.dwarf2/dw2-inline-with-lexical-scope.c (func): Replace
	a dead code with an assignment to a global var.  Fix a
	whitespacing problem around an assignment operator.